More structural setup
For GTU, was processing a raw tra file into a series of STRING_SETs. This has been generalized into a function so it can be used
in core fixes. Also added a function to process spell.ids into a series of variables named after the spells. Added the Symbol,
Pain description fix from IWDification/SCS as a test case.