[bg, bg2, iwd] Fixing issue with cure spells and trolls
https://github.com/Gibberlings3/EE_Fixpack/issues/28

Pushing the cure sleep (2) opcode into a standalone cure spell delays it juuuuust enough in the effect stack so that it removes
the sleep effect used to simulate death. The various troll death spells on all platforms now cure sleep directly and then block
further applications for a second to allow this to work normally. Bit of a hack, but it works and doesn't mean modders have
to account for these troll spells on their own.