[bg, bg2, iwd] Spells-via-trap can crash the game
Originally reported by JediMindTrix on Discord, verified as a code issue by Bubb, and filed by argent77 here:

https://github.com/Gibberlings3/EE_Fixpack/issues/35

Essentially, when a 324 is used the engine looks for a source to use in the immune feedback string. If it's a spell with no
name and it's being invoked by  a script (e.g. a trap) the game will crash. This will occur consistently with a 324, but can
also happen when using a 318. While this should be fixed at a code level, we can work around it simply by naming spells used in
this fashion. I've gone through and identified all spells being cast by trap scripts and reviewed them for these two conditions.
Any spell that qualifies is simply named 'Trap' so that there is a valid feedback string for the 318/324.