[bg2] Fixing eye graphics in SoA abyss finale
https://www.gibberlings3.net/forums/topic/34875-bug-report-ee-content-bugs/page/8/#comment-314950

Two fixes in one: running the graphic through a VVC to fix blend and placement. Noticed during testing that it was nearly
impossible to see the topmost eye animation, so made the main door no longer block LoS to make the graphic visible to the
player instead of the eye simply closing.