General speed improvements
The easy one is wrapping the ALWAYS stuff in a variable check so we only execute it once per install.

The longer one is the tra > STRING_SET process. While generating and executing STRING_SETs from an arbitrary tra file is neat,
it's also slow. I've altered the function and now pre-generate tph files for inclusion by the mainline and GTU components,
rather than having end users wait on the process. This will need to be run prior to each release.