[bg, bg2, iwd] holy smite/unholy blight fixes
On BG/BG2, Holy Smite and Unholy Blight are changed to use 324s instead of a horrific pile of EFFs.

On IWD, the AC penalty for Unholy Blight is changed to be a -2 to damage rolls, per the descript.

On BG/BG2/IWD, Unholy Blight now uses a custom icon for the duration of its penalties.

Also fixed a minor issue preventing install on BGEE without SoD.