[bg] Aldeth's doppleganger quest
These dopplegangers just don't like to be counted properly, so we're going overkill. counting is now handled by the area script, and
should account for both pre- and post-transformation dopplegangers. originally included the exterior of the building, but I worried about
double-counting since it's a master area script. Since the only places these are meaningfully checked are in the sub-areas, I'm not too
concerned about excluding the exterior.