[bg, bg2, iwd] More minor spell fixes
https://www.gibberlings3.net/forums/topic/34875-bug-report-ee-content-bugs/page/16/#findComment-331726
https://www.gibberlings3.net/forums/topic/34875-bug-report-ee-content-bugs/page/16/#findComment-331734
https://www.gibberlings3.net/forums/topic/34875-bug-report-ee-content-bugs/page/16/#findComment-331740

Did pretty much what's in the links; a few had been addressed previously in other patches. The Protection From Evil 10' changes
necessitated a lot of text updates, as the string needed to be updated plus it was already in a number of GTUs. Also noted (and fixed)
the BGEE scroll of same using different strings than the underlying spell.