[bg, bg2] portrait.2da defaults
https://github.com/Gibberlings3/EE_Fixpack/issues/81

Many player-selectable portraits didn't have entries in portrait.2da. Generally this isn't an issue--the primary function of portrait.2da
is to give recruitable NPCs an alternative portrait if the player selects their portrait for their PC. However, this does cause issues in
the Black Pits, as selecting multiple characters with the same portrait will force the subsequent ones onto the alternative in
portrait.2da. If there's no listing they get man2s.

This will provide generic defaults for all of the portraits that don't have an entry.