[bg, bg2, iwd] bulk cre fixes + golem class fixes
Two items in the commit:
* Moved the race/class updates from the IDS sync into the bulk creature fixes.
* Added a few class updates from subtledoctor in this thread https://www.gibberlings3.net/forums/topic/39319-golems/

The ice golem fixes were already in as part of the IDS sync, but added the bone golem and greater clay golem updates, though I'm omitting
magic golems for now. The suggestion from SD was to use OGRE_MAGE since nishruu/hakeshar use it, but both of those have been moved to MIST
so there's not a natural class to assign. We may need to add one for them though we're down to ~30 available class slots.