language changes
MephistoSatanDevil submitted some Chinese-language GTU stuff, but since they're actually crash fixes I've moved them into the main fix
component proper, and tph files have been generated for them. The contents of the English fixes_(game).tra files should probably be reviewed
by the Chinese translation community for whether or not they are needed as well.

Also, many languages have native text but English voiceovers. This was handled on a one-off basis in the IWDEE soundset fixes, but I've
updated this to be handled globally with a new variable, vo_lang, set in the ALWAYS block. The soundset fix is updated to use this
instead, and the IWDEE chapter six narration fix is now keyed off of it as well.