[bg2] brannel dialogue triggers
Artisan on Discord:

A very minor issue but I think it could be something to fix. In the thief stronghold questline the doorman Brannel is supposed to start a
dialogue and give you a hint for the Ana betrayal, but if you ignored him which most players will likely do then he brings up his
introduction dialogue instead as soon as you leave the building and you have to talk to him again making the warning very missable.

You could either have the warning take priority or give the doorman a script to force dialogue when you approach the first time around,
since that dialogue is written like like it should be NPC initiated anyway.